+++ frontend/vitest.config.ts
@@ -1,0 +1,22 @@
+import { defineConfig } from "vite";
+import { configDefaults } from "vitest/config";
+import react from "@vitejs/plugin-react-swc";
+
+export default defineConfig({
+  plugins: [react()],
+  test: {
+    globals: true,
+    include: ["src/**/*.{test,spec}.?(c|m)[jt]s?(x)"],
+    environment: "jsdom",
+    setupFiles: "./setupTests.ts",
+    coverage: {
+      exclude: [...configDefaults.coverage.exclude, "src/index.tsx"],
+      thresholds: {
+        lines: 80,
+        functions: 80,
+        branches: 80,
+        statements: 80,
+      },
+    },
+  },
+});